% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_ml_irm.R
\docType{class}
\name{DoubleMLIRM}
\alias{DoubleMLIRM}
\title{DoubleMLIRM R6 class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
DoubleMLIRM
}
\description{
DoubleMLIRM R6 class
}
\keyword{datasets}
