% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_irmiv.R
\name{orth_irmiv_dml}
\alias{orth_irmiv_dml}
\title{Orthogonalized Estimation of Coefficient in irm}
\usage{
orth_irmiv_dml(p_hat, mu0_hat, mu1_hat, m0_hat, m1_hat, d, y, z, inf_model)
}
\arguments{
\item{d}{Treatment variable.}

\item{y}{Outcome variable.}

\item{inf_model}{Estimator for final estimation, default average treatment effect \code{"ATE"}. Alternatively switch to \code{"ATET"} for average treatment effect on the treated.}
}
\value{
List with estimate (\code{theta}).
}
\description{
Function to estimate the structural parameter in an interactive regression model (irm).
}
